+++ app/Http/Controllers/Api/DepartmentController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Api\Concerns\HandlesRoleAccess;
 use App\Http\Controllers\Controller;
 use App\Models\Department;
 use Illuminate\Http\Request;
 
 class DepartmentController extends Controller
 {
+    use HandlesRoleAccess;
+
     public function index()
     {
         return response()->json(Department::all());
@@ -15,6 +18,8 @@ class DepartmentController extends Controller
 
     public function store(Request $request)
     {
+        abort_unless($this->canManageDepartments($request->user()), 403, 'You are not allowed to create departments.');
+
         $validated = $request->validate([
             'department_name' => 'required|string',
             'department_code' => 'nullable|string|max:20',
@@ -27,6 +32,8 @@ class DepartmentController extends Controller
 
     public function update(Request $request, $id)
     {
+        abort_unless($this->canManageDepartments($request->user()), 403, 'You are not allowed to update departments.');
+
         $department = Department::findOrFail($id);
 
         $validated = $request->validate([
@@ -41,6 +48,8 @@ class DepartmentController extends Controller
 
     public function destroy($id)
     {
+        abort_unless($this->canManageDepartments(request()->user()), 403, 'You are not allowed to delete departments.');
+
         $department = Department::findOrFail($id);
         $department->delete();
